Redesigned Picture image link generation. Instead of getting a temporary key for Azure storage containers, images can now be loaded directly from /Pictures/File?id={PictureId}&size={Size of the image}.
For Mobile clients the Azure storage container links are still needed, as I can't update the mobile app right now.

Added an extension method to Pictures for easy construction of urls to link to the image file.

Added File(int id, int size) action to Pictures controllers in both Web and ProgenyApi.
This returns a FileContentResult with the image.

Updated link generation in Web PicturesController, HomeController, and TimeLineItemService to use the new link generation method.