Replaced hard coded translations in IDP with API calls to the Translation API. Files updated: - IDP/Controllers/AccountController.cs - IDP/Services/IEmailSender.cs - IDP/Services/EmailSender.cs
Made LocaleManager service's GetAllLanguages() method public and added it to the interface to be able to fetch a list of languages from other parts of the code.

Added integer language Id property to ChangePassword, ForgotPassword, and Register pages' view models and forms, to directly set translation language in the AccountController actions.