Add background task management and sample picture extension check
Introduced `KinaUnaBackgroundTask` entity and added a new `DbSet<KinaUnaBackgroundTask>`  to `ProgenyDbContext` for managing background tasks.

Updated  created a migration for updating the database.

Added `CheckPicturesForExtensions` method to `IPicturesService` and implemented it in `PicturesService`.

Modified `UpdatePictureLinkWithExtension` to handle exceptions.

Registered `IBackgroundTasksService` and `IRunTaskService` in `Startup`.

Updated `ITasksHttpClient` and `TasksHttpClient` to include methods for resetting and executing tasks with `KinaUnaBackgroundTask` parameter.

Added `BackgroundTasksController` and `RunTasksController` for managing and running background tasks.

Implemented `BackgroundTasksService` and `RunTaskService` for task management and execution.