Add support for UId property in calendar items
Added UId property copying in CopyPropertiesForUpdate and CopyPropertiesForAdd methods in CalendarItemExtensions.cs.

Added logic to generate a new UId if it is null or whitespace in AddCalendarItem and UpdateCalendarItem methods in CalendarService.cs.

Added UId property assignment in CalendarItemViewModel class and its ToCalendarItem method in CalendarItemViewModel.cs.

Included hidden input for UId in _CopyEventPartial.cshtml and _EditEventPartial.cshtml.