Add InheritPermissions to TimelineItemPermission entity
This commit introduces a new boolean property `InheritPermissions` to the `TimelineItemPermission` entity in the database schema. The property is added to the `TimelineItemPermissionsDb` table, enabling the specification of inherited permissions. Updates to the `ProgenyDbContextModelSnapshot.cs` file reflect this change, ensuring Entity Framework Core migrations recognize the new schema. A new migration class `AddInheritToTimelineItemPermission` is also created to manage the addition of this column, including methods for applying and reverting the migration.